pyproject.toml
[project]
name = "documind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.11.0",
    "celery==5.4.0",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "channels==4.3.2",
    "channels-redis==4.3.0",
    "charset-normalizer==3.4.4",
    "cryptography==46.0.3",
    "dj-rest-auth==7.0.1",
    "django==5.2.9",
    "django-allauth==65.13.1",
    "django-cors-headers==4.9.0",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "drf-yasg==1.21.11",
    "idna==3.11",
    "inflection==0.5.1",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-postgres>=0.0.9",
    "langchain-text-splitters>=0.3.0",
    "msgpack==1.1.2",
    "openai>=1.50.0",
    "packaging==25.0",
    "pillow==12.0.0",
    "psycopg2-binary==2.9.11",
    "psycopg[binary]>=3.1.0",
    "pycparser==2.23",
    "pyjwt==2.10.1",
    "pypdf>=5.0.0",
    "python-decouple==3.8",
    "python-dotenv==1.2.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "redis==7.1.0",
    "requests==2.32.5",
    "sqlparse==0.5.4",
    "tiktoken>=0.7.0",
    "tzdata==2025.3",
    "uritemplate==4.2.0",
    "urllib3==2.6.2",
]
